requirements.txt
Django~=4.2.1
asgiref==3.6.0
djangorestframework==3.14.0
Pygments==2.15.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
django-rest-registration==0.8.1
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.6.0
Markdown==3.4.3
PyJWT==2.7.0
zipp==3.15.0
python-dotenv~=1.0.0